requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
runwayml==3.9.0
aiofiles==23.2.1
openai==1.3.0
pytest==7.4.3
pytest-asyncio==0.21.1 